refactor: split coca implementation into multiple files
