+++ tests/models/coca/coca_config.yaml
@@ -8,7 +8,7 @@ vision_encoder_config:
   prediction_key: vision_embeddings
   block_size: 196
   n_classes: Null # Disable vision transformer head
-  n_layer: 12
+  n_layer: 6
   n_head: 8
   n_embd: 768
   dropout: 0.0
@@ -20,25 +20,6 @@ vision_encoder_config:
 text_decoder_config:
   sample_key: input_ids
   prediction_key: text_embeddings
-  block_size: 1025 # +1 for the class token
-  vocab_size: 50304
-  n_layer: 12
-  n_head: 12
-  ffn_hidden: 2048
-  n_embd: 768
-  dropout: 0.0
-  bias: true
-  attention:
-    attention_type: pytorch_flash_attention
-    scaling_factor: 3
-  activation: fused_swiglu
-  epsilon: 1e-5
-  weight_init:
-    mean: 0.0
-    std: 0.02
-multimodal_decoder_config:
-  sample_key: text_embeddings
-  prediction_key: logits
   block_size: 1024
   vocab_size: 50304
   n_layer: 12
@@ -59,3 +40,4 @@ n_pool_head: 8
 n_vision_queries: 256
 bias_attn_pool: False
 epsilon_attn_pool: 1e-5
+n_shared_layer: 6
