+++ tests/models/coca/test_coca.py
@@ -4,7 +4,7 @@ import torch
 
 from modalities.__main__ import Main, load_app_config_dict
 from modalities.config.config import AppConfig
-from modalities.models.coca.coca_model import AttentionalPooling, CoCa, CoCaConfig
+from modalities.models.coca.coca_model import CoCa, CoCaConfig
 from tests.conftest import _ROOT_DIR
 
 
@@ -18,7 +18,7 @@ def test_coca():
     # Create dummy inputs
     dummy_input_image = torch.randn(1, 3, 224, 224)
     dummy_input_text = torch.randint(
-        0, coca_config.text_decoder_config.vocab_size, (1, coca_config.multimodal_decoder_config.block_size)
+        0, coca_config.text_decoder_config.vocab_size, (1, coca_config.text_decoder_config.block_size)
     )
     dummy_input = dict(images=dummy_input_image, input_ids=dummy_input_text)
 
@@ -39,14 +39,6 @@ def test_coca():
     assert out["logits"].shape == (1, 1024, 50304)
     assert out["vision_cls"].shape == (1, 1, 768)
     assert out["text_cls"].shape == (1, 1, 768)
-
-
-def test_attn_pool():
-    model = AttentionalPooling(n_embd=768, n_head=8, bias=False, epsilon=1e-5)
-    dummy_vision_embed = torch.randn(1, 256, 768)
-    dummy_queries = torch.randn(1, 257, 768)
-    out = model(dummy_vision_embed, dummy_queries)
-    assert out.shape == (1, 257, 768)
 
 
 def test_e2e_coca_training_run_without_checkpoint(monkeypatch):
